Optionally clean up the conda-build build environment
This build environment was created by older versions of `conda-build`,
but has since been forgone in favor of using an environment that fills
the full path length used by binaries. However, with newer versions of
`conda` if the environment does not exist then removal is an error. Here
we make a change so that this potential error does not derail the build,
but merely warns about the absence of the environment.