Optionally clean up the conda-build test environment
The `conda-build` test environment is not created as we don't have a
test step. Previously `conda` would silently fail with missing
environments. However it seems the new behavior is to error out if the
environment does not exist. The change here suppresses the error so that
it does not fail the build. After all we will still want the ability to
remove this environment if it were to exist.