Support base64 encoded traces
Bokeh 0.12.4 automatically encodes NumPy arrays to base64. The resulting
decoded arrays in JavaScript are 1D. However, for the traces we need to
be able to operate with it as if it was a 2D array. Here we make some
adjustments to handle the shape and striding manually so as to display
the traces from this more dense format.

We also design this strategy so that it will continue to yield a
functioning strategy on older versions of Bokeh (e.g. pre-0.12.4). This
way if there is no support for base64 encoded traces, the plots can
still be generated and work fine. Simply if Bokeh 0.12.4 is present, we
benefit from all the optimizations it has to offer.