Use the `LinearColorMapper` for greys
Instead of supplying a palette to Bokeh, which seems to result in a
`LinearColorMapper` being generated anyways with two copies of the same
color range being embedded in the plots, just create a
`LinearColorMapper` to pass to the intensity plots. This seems to result
in only one copy of the colors being embedded in the plot and the same
appearance.