Copy ROI masks in chunks
This makes a change to how ROI masks are copied from the postprocessing
results into the ROI data. Instead of loading all of the ROI masks into
memory and copying them over, load a group of masks that is less than or
equal to the block frames and copy that to the ROI data. This works well
in a number of cases. If the number of ROIs is less than the number of
block frames, then this amounts to the same behavior of copying all the
ROI masks over.