Use black background for traces
As some of the ROIs have very bright colors (e.g. yellow) that don't
show up very well on a black background, it seems black might be a
better choice. So this changes the trace background color to black to
help all the traces to pop and make it easier to identify what is going
on with them.