Write out our own custom HTML file
Instead of using Bokeh's mechanism for writing out the HTML file of the
plots, roll our own. The main reason for doing this is to get access to
elements outside the scope of Bokeh. Here we borrow heavily from Bokeh's
API to fill out the HTML file. Though largely this file appears the same
as it did when Bokeh generated it.