Make plot outlines white
While the axes showed up ok after they were set to white, the opposing
edges of the plot still looked a bit thin. We fix this by setting the
plot outlines to white as well, which fixes this. This change is not
immediately apparent, but can be verified to be working by comparing
plots with and without this change.