Display the Bokeh HTML file in the notebook
Instead of displaying Bokeh's Notebook output, display the HTML file
that we wrote out to disk in the Jupyter Notebook. This allows us to
view the file as it was written and with any modifications that are
relevant. This way things like other formatting changes that are out of
the scope of Bokeh still show up. Also any dependencies that we add will
also be available within the page. Overall this should also reduce the
differences between what is displayed in the Jupyter Notebook and the
HTML file that one views later.

We need to use an `iframe` when displaying the HTML instead of including
the full HTML as Jupyter Notebooks do not handle the latter. This likely
has to do with some formatting they do in the background to make sure it
fits within the cell and the page overall. We have also found this is
necessary to make sure external JavaScript dependencies work correctly
(e.g. BokehJS and others).

The `iframe` is formatted to allow the width to expand to the full
width of the cell (whatever that might be). The height is determined
based on the plot size with an approximate formulation that should be
sufficient for fitting both plots. In general, these settings should
minimize the need for scrolling in the `iframe`. Particularly it should
avoid scrolling in the vertical direction. Also it should make it easy
to view the traces and ROIs in the same field of view given the default
settings. However, scrolling within the `iframe` may still be necessary
if the plots are resized.