Store traces as `uint8`
To prepare for working with compression of the trace data, this converts
the trace data to `uint8`. This makes for a good format for both storing
the compressed trace data in the HTML file and decompressing and
converting the trace data back to usable form in JavaScript.

In order to make conversion go smoothly, add a dummy singleton 1D array
of the same type as the trace data to the callback. This allows the type
information of the trace data to be properly converted by Bokeh to its
expected `TypeArray` in JavaScript. Using the type of the dummy array,
construct a view of this type onto the trace data's buffer. This avoids
any unnecessary copying, but ensures that the trace data can be treated
using the proper type.