Store decoded traces
Instead of decoding the traces every time the selection callback is run,
decode the traces once and overwrite the callback with the decoded
value. Also note that decoding has already occurred by setting the dummy
singleton array value to `1`. As every event that triggers the callback
checks against this dummy value, we can ensure the decoding will only
happen the first time.