Require Bokeh 0.12.4+
Already we benefit from optimizations provided by Bokeh 0.12.4. Namely
the `base64` encoding of arrays makes our trace plot HTML files much
more compact. We had added some code to ensure things continued to work
with Bokeh 0.12.3 while there was not a formal release of Bokeh 0.12.4,
but would leverage the features of the newer Bokeh if available.

However, now that Bokeh 0.12.4 is out and 0.12.5 is on its way, it
doesn't make sense to continue to support the older Bokeh. Upgrading to
the newest Bokeh is a simple matter. We already ship Docker images with
0.12.4. As we work seamlessly with Bokeh 0.12.4 and it generates much
more compact plots for us, there seems to be little benefit to
supporting an older Bokeh.

In addition, to make further optimizations to the HTML file size, we
will need to explore compression. This requires that we leverage type
information of the data displayed. It is not possible to do this in
Bokeh pre-0.12.4 as everything is treated as generic `Array`s of
`Number`s. Instead we will need to require Bokeh 0.12.4+ in order to get
this type information.

For both of these reasons, we change our Bokeh requirement to have a
lower bound of 0.12.4. This way anyone upgrading the workflow will get a
copy of Bokeh that supports these new features.