Seed initial dictionary per atom for all types
Rewritten so as to handle NumPy, HDF5, Zarr, and LazyDataset cases.
NumPy and LazyDataset are able to handle the slice without help in this
case. HDF5 and Zarr are not, but both return a `TypeError` when they
fail. So instead of trying to ask for permission we ask for forgiveness
and handle the `TypeError` case specially.