Store post-processing results in an HDF5 group
Instead of storing the full structured array as one compound HDF5
dataset, create a group and store each record as a separate dataset.
This allows us to more easily extract the ROI masks in the next step. It
also works nicely with the current version of Zarr and HDFView. Not to
mention it is just easier to work with.