Readd the `LazyDataset` as a parent class
This provides us a base class to work with, inherit from, and refactor
some common functionality into. Should help us provide something similar
for Zarr as currently exists for HDF5.

Note: Would have used an abstract base class, but it appears these are
not picklable. So have opted to make it a normal class with some dummy
methods.