Place results in Zarr before transmission
This should cutdown on the size of results computed in parallel before
transmitting them back. We are able to do this by compressing the data
first. While we haven't found this to be a terribly bad bottleneck ATM,
it certainly would be nice to leverage Zarr to its full potential. Also
it should help in situations where the network is near saturation.