Add a context manager to get a Distributed Client
Provides a handy utility function to bootstrap a Distributed Client on
top of an ipyparallel cluster. As it is a context manager, this
guarantees that the Distributed Cluster can be cleaned up at the end of
the `with` block. Given that it doesn't seem very easy to cleanup a
Distributed Cluster when not already having access to it (unlike with an
ipyparallel cluster), handling the shutdown and cleanup of the
Distributed Cluster can be done easily at the end of the block.