Add function to map computations over a Dask
The utility function blocks a Dask up into small haloed Dasks. Once
broken up, it maps a computation over each haloed block and trims the
halos. With the resulting blocks, it then concatenates them back into
one Dask that contains the result. However, as we are use to working
with blocks that are a specified size, a flag is provided (defaulting to
true) that allows converting this result into blocks to allow for
handling storage while tracking progress.