Change git's clean filter to ignore cell metadata
Simply adjust the `clean_nb` filter in `.gitconfig` to set each cell's
`metadata` field to an empty dictionary. This is the same behavior the
notebook server has when saving an empty notebook. So this consistent
with the existing behavior of the filter to treat the notebook as if it
hasn't been run. Also the `metadata` field only defines things like how
to display the output from running the cell (e.g. is it scrolled, should
it be collapsed, etc.). So it is good to leave this undefined initially.
That way the user can make these choices when running themselves.