Specify basename for data files
Instead of guessing what the basename for all data files should be based
on the input data, just let the user specify this. Update all operations
to simply use a postfix for specifying their particular type of output
data as different from the rest. This should make things more flexible
as we consider different forms of input data.