Support lazy loading of TIFFs with Dask using PIMS
Adds a function to `nanshe_workflow.data` that allows for fast lazy
conversion of TIFF images into Dask Arrays using the PIMS library to
handle metadata inspection and actual reading of frames. This allows us
to handle the TIFFs like one big array without taking a significant
penalty because of their size or the fact that they may not fit in
memory.