Add function to lazily load HDF5 with Dask
In most cases, one can use an HDF5 Dataset like a large lazily loaded
array. However working HDF5 files in parallel can be a bit tricky. In
Python, one cannot pickle up an HDF5 file and pass it between processes.
This goes from dangerous if it is a file open for writing to simply
unsupportable when the file is read only. That said, one can open the
same file for reading multiple times without issues. Still this becomes
quite a bit of bookkeeping for each parallel process to handle.

However, we can improve on this significantly if we leverage Dask
Delayed. What we can do is create a function to open an HDF5 file, grab
the Dataset, and slice out a chunk as a NumPy Array to return. We then
wrap up this function with Dask Delayed so that we can delay the actual
read and so that we can convert the Dask Delayed object into a Dask
Array. With the hard part done, it is a simple matter to iterate over
all chunks in the HDF5 Dataset this way. Once we have all the chunks as
Dask Arrays, we can then stack them into one Dask Array representative
of the full HDF5 Dataset.

With the HDF5 Dataset as a Dask Array, we are freed up to do anything we
would want to do in parallel with an immutable HDF5 Dataset. Also the
bookkeeping involved in reading each block is now handled behind the
scenes in our Dask Array. So different processes don't need to get
involved in handling this bookkeeping. As this strategy only tries to
pickle things like strings, we don't need to get involved in trying to
handle passing HDF5 files around between process. In short, we get all
the benefits of HDF5 as a storage format without requiring Dask know too
much about it.