Convert HDF5 Dataset to Zarr using Distributed
Leveraging the fact that we can now ingest an HDF5 Dataset into Dask,
load the input HDF5 Dataset this way. As it is now a Dask Array, we can
handle storing to Zarr through standard Dask mechanism. The result is
handling conversion of an HDF5 Dataset is not so different from handling
TIFF files.