Handle singleton slices when ingesting TIFFs
Updates the code in `dask_imread` to use a singleton slice when indexing
the image. This is mainly demonstrate what it would look like to read
data this way. It also helps better isolate the changes required for
non-trivial `slice`s. Finally this is can help compare the run time of
the two approaches to chunking TIFFs in Dask per frame.