Support chunking of Dask-loaded HDF5 and TIFFs
As chunks were already supported with HDF5, it is a simple matter to
allow the user to specify these externally. TIFFs require a little bit
more rearrangement to handle chunking along frames. However, this is
still doable. Skipped chunking along other dimensions as it seems to
have unclear value ATM.