Add a simple denoising step
Given we seem to need to do some denoising before performing
registration and it is helpful beyond the registration step, it makes
sense to make this its own step. Basically perform a simple median
filter over each frame to remove speckle noise and then subtract the
convolution with a large Gaussian to remove light effects.