Rechunk denoising's smoothed FFT to match input
Instead of rechunking the smoothed FFT to have a single frame and full
shape specifically. Simply rechunk the smoothed FFT to have the same
chunks as its input data has. This seems to be important on larger data
where doing a full rechunking from scratch eats up a fair bit of time
both in graph construction and computation. By doing this instead, we
merely fuse the chunks along the shape dimensions, which seems to be
much faster on large data.