Use `map` to run `chdir` on workers
To make sure that all workers change to the data directory, use `map`
and sequences that are of the length of the number of workers. This is
done to make sure all of them change and not just some of them. Also
have workers change to the data directory not just the current directory
of the client (even though these should be the same).