Compute denoising's FFT inverse before subtraction
In an attempt to speed up the denoising step a bit, simply subtract the
smoothed frames after performing the inverse FFT on them. This way Dask
won't need to keep the smoothed and unsmoothed FFTs in memory. Instead
it will only need to keep the median filtered data and the smoothed
data to perform the subtraction. Thus it should lighten the memory
requirements and speed up the computation.