Enable parallel writing with Zarr
Allow chunks to be written in parallel by Zarr. As we already guarantee
that the final Dask result will have the same chunk size as the dataset
of the on disk Zarr, we don't need to worry about collisions between
chunks or multiple writes to a single chunk. Instead every chunk that is
written out from Dask has a one-to-one correspondence with every chunk
in Zarr. So this writing is embarrassingly parallel and can be done
trivially. Thus we dispatch with the locks as they are unneeded in this
configuration.