Positively shift results from denoising
The subtraction of the convolution with a normalized Gaussian can result
in negative values in the image, which is undesirable. The reason being
the negative values propagate into later steps and can result in ROIs
being missed or appearing smaller than they should. To address this, we
simply shift the data so as to have the same minimum that it started
with. However the maximum value may be different.