Shift the harmonic mean subtracted result positive
Fix an issue with the harmonic mean projection subtraction where some
negative values would appear. This causes some issues with later
computations that may miss ROIs or draw them to small as a consequence.
To fix this, the step had to be rewritten with Dask. However that made
the code simpler as well. Also simply subtract the minimum from the
final result so that the minimum value in the data should be zero.
As the point of this step is to remove any mostly unchanging background
in the image, it makes sense for the minimum to be fixed at zero in this
step. This doesn't have any adverse effect on later steps.