Add a step for trimming the edge frames
Often in real world data, we encounter that the initial frames from the
experiment just need to be tossed. This could be due to the camera
getting bumped, bad settings on the camera, missing parts of the first
frame, etc. Similarly issues can show up in the last few frames for
somewhat different reasons. As these frames end up being so bad, so few,
and often well outside the experimentally relevant time period, it ends
up being just easier to toss them than trying to do anything else with
them. Hence we are adding a step to the workflow that allows exactly
this sort of cleanup. The user can specify how many frames from the
beginning and end that they would like to remove. By default, it removes
none, but the option exists should it be needed.