Drop informal frame trimming from denoising
As we have a formal trimming step before denoising, there is no need for
an informal trimming in denoising where we remove a frame on either end.
Instead we can just handle this in the trimming step and denoising can
just be focused on trying to remove persistent noise before proceeding
to later computations.