Use Real FFTs in Image Registration
In order to speed up image registration a bit, make use of Real FFTs
both to convert individual frames to frequency space and to find the
convolution of frames with the template image in real space. Given the
images and the result of the convolution should be real, these are
reasonable constraints to make.