Drop explicit extraction of real FFT results
In the code, there were leftover calls to get the real part of the
result from `irfftn`. However `irfftn` already returns a real result. So
there is no imaginary part to strip off. As such, we can drop these
requests for the real part of the result as it is already satisfied.