Perform Gaussian filter in real space
It seems that performing the Gaussian filter in frequency space results
in the computation being blocked after partial completion. It's unclear
what the cause for this is, but it may be some inefficiency in how that
computation proceeds. Though it could also be highlighting some
underlying bug in Dask. For now, we switch to applying the Gaussian
filter in real space, which doesn't encounter this issue. Also this
allows for other spatial chunk sizes to be explored, which can be used
to tune this step if it is slow.