Add a border to the test data
In order to ensure registration is properly tested, we need to run a
non-trivial number of iterations with it. However this causes us a
problem on the existing dataset as there is no background to perform the
registration against. To fix this issue, we add an elevated boundary to
all frames. This gets removed as part of background subtraction.  Though
this still has an impact on registration where it encourages aligning
the boundary across frames. Hence this blocks registering against
non-stationary elements.