Require dask-core (instead of dask)
When it comes to using Dask, we are interested in using it with NumPy
arrays on a Distributed cluster (bootstrapped off of ipyparallel).
However we don't really use Pandas or any tabular data (outside of NumPy
structured arrays, which we are using less of). As such, pulling in
everything needed to run Dask is a bit much. It makes more sense to take
an a la carte approach to Dask and simply install those dependencies
with it that we wish to use (e.g. NumPy and Distributed). Hence we
switch to the `dask-core` package instead of the `dask` package, to
allow this flexibility and reduced requirement load.