Add blocks space parameter to normalization step
The `block_space` parameter was being used in the data normalization
step.  However it was not being set in that step. Instead it was being
carried over from a previous computational step. This adds `block_space`
as a parameter for the normalization step and documents it as one.
Should make things a bit clearer.