Use ZipStore for copying Zarr directly
When using Zarr's `open_group` with a `ZipStore`, it creates `.zattrs`
and `.zgroup` to start with. However we want to copy in our own
`.zattrs` and `.zgroup` from a different Zarr (presumably using a
`DirectoryStore`). As a result, we end up creating the `.zattrs` and
`.zgroup` entries in the Zip file twice. While the newer ones copied
over do overwrite the older ones, it would be better if we just avoided
creating these entries the first time.

This change tries to avoid creating `.zattrs` and `.zgroup` to start
with by simply opening the `ZipStore` directly and then copying content
in from the Zarr provided. By doing this, we can ensure the Zip file has
no entries whatsoever and merely gets the entries that we copy over from
the newer Zarr. Hence there shouldn't be any `.zattrs` or `.zgroup` to
start with and we shouldn't get any warnings about duplicating these.