Revert "Flush after writing each Zarr ZipStore entry"
This reverts commit a450c82642a2d80574b73ad78e07ef725a2dfdfe as it was
found that flushing after each entry was way too slow. Remains unclear
whether this would help with the one off memory issue that we saw.
Based on looking at the CPython codebase, it seems each value is written
out immediately and locking is done around the Zip file (as parallel
writing is not supported/unsafe). So it seems unlikely that this could
have caused the issue anyways.

Given the performance issues and unclear benefits, flushing after each
entry doesn't seem like a penalty that is necessary to take without more
evidence. As such the codebase is being returned to a state where the
`MutableMapping` method `update` is used instead. This should return the
nice performance that was seen previously.