Generalize each registration's shift's shape
Previously, in our translation-based image registration algorithm, we
had hard-coded the shape of each shift computed to be based on 2-D
frames. However one could conceivably use this algorithm on N-D frames.
Though the hard-coding here blocked using this on some higher or lower
dimensional data. So fix the shape of each shift to be determined by the
dimensionality of each frame checked for translation instead as this
ultimately affects the shape of each shift.