Find possible shifts in Dask
Instead of passing off shift determination to a Dask `delayed` function,
which will have to load the full frame for the computation, handle
finding the shifts in Dask directly. Also wrap large shifts to get
smaller shifts in the opposite direction (using our periodic boundary
conditions). However we still cannot find the minimum of an array with
unknown length. So continue finding the best shift through a Dask
`delayed` function. The rest of registration proceeds the same after
acquiring the best shifts for each frame.