Add directory zipping function
Provides a simple utility function for zipping up the contents of a
directory. Skips any empty directories only adding files to the archive.
Proceeds topdown through the directory and keeps all filenames in a
sorted order. The archive is simply named after the directory with
`.zip` appended to it. Returns the archive location to the user as a
convenience.