Use directory zipping function for zipping Zarrs
Instead of opening the `DirectoryStore` with Zarr to zip it, simply scan
through the directory's contents with `os.walk` and add each to a
`ZipFile` directly using the `write` method. This method ensures that
each file is copied over in blocks of a fixed size (or smaller). So
should help keep memory usage within certain bounds when creating the
Zip file.