Get the N-D shift per frame in registration
When determining how much to shift each frame in registration, retrieve
the full N-D shift for each frame instead of retrieving the shift per
dimension. As the shifts should already be chunked per frame, it should
be more efficient to simply get the full shift for a single frame than
trying to grab each dimension component of the shift for each frame.