Fuse best shift computation in registration
Instead of iterating over each frame and determining the best shift,
this resurrects `compute_offset` and makes a few tweaks so that it can
find the best shifts for a small selection of frames. Using this
modified version of `compute_offset`, this now runs `atop` on the
matches found.  Note that Dask's `delayed` is now removed as a
consequence. This should result in significantly fewer objects for Dask
to track during graph construction. Also removal of Dask `delayed`
should avoid some costly optimization passes on the Dask graph. Both of
these changes should have a notable impact on graph construction time
and complexity for large videos.