Use scandir's walk
Instead of using `os.walk`, use `scandir.walk` for better performance on
Python 2. As we are currently using Python 3.6, this has no impact on
that Python version since its `os.walk` implementation is the same as
`scandir.walk`'s. Though if we have to go back to Python 2 because of
some dependency, this should make sure the performance does not suffer
in this area.