After Dask Distributed progress bar print newline
Make sure to print a newline right after the Dask Distributed text
progress bar completes in the registration step. This is needed as the
text progress bars will write over themselves. Also any text that is
printed after the text progress bar completes, will get shoved into the
same line as the progress bar. While we certainly had worked around this
issue by printing right before we printed any text, this makes it more
clear that this is needed as a workaround for the text progress bar's
behavior.