Optionally use parallel removal in `zip_zarr`
Make use of `dask_io_remove` (our parallelized version of `io_remove`)
in `zip_zarr` if the user provides the `executor` argument with a Dask
Distributed cluster. Otherwise falls back to vanilla `io_remove`, which
is also the default. Should make it easy to convert Zarr files to the
Zip-based format fast.