Delete Zarr directories in parallel
While we cannot actually write to Zip files in parallel, we can delete
the old Zarr directories in parallel. This change parallelizes the
removal of all old Zarr directories after the Zarr has been stored in a
Zip file. Leverages the Dask Distributed cluster to perform this
parallel removal. Should speed up removal of these directories on NFS
where this is very slow when single threaded.