Make `dask_rm_dir` robust to a non-empty directory
In case removal of the directories contents does not proceed as
intended, make sure to delete the directories contents in the directory
removal task. This really shouldn't ever happen, but sometimes NFS can
be a little slow about propagating changes to all of the nodes. This
simply makes sure that we make the change ourselves if needed.