Quietly handle removal of absent file or directory
If the file or directory to be removed does not appear to be there in
our task, catch this case after the removal attempt and suppress the
error. However if the error is anything other than a missing file or
directory, re-raise it so that it can be handled more appropriately.
Design this approach with Python 2/3 compatibility in mind.