Use `dask_store_zarr` throughout the workflow
Instead of having a bunch of duplicated code, which is a bit tricky to
parse, for storing results to Zarr from Dask Arrays in the workflow,
simply use the refactored function, `dask_store_zarr`, to store any
results after designing the computation for each step. This makes the
workflow much easier to follow, should avoid mistakes related to
copy-pasta of this code, and should make it easier for users to add new
steps based on these ones.