Use Dask to find min/max of the data to view
Instead of using old ipyparallel-based array functions to find the
minimum and maximum of each image, just use Dask and the Distributed
cluster to compute the minimum and maximum of each image. This should
even be a bit better memory-wise for both workers (during the reduction
stage) and the client as they will be passed single values instead of
taking the maximum of array slices.