Extract ROIs and compute label image with Dask
Makes use of Dask to extract the ROIs from post-processing and convert
them for use with ImageJ and MATLAB. Also constructs the label image
using Dask operations. This should be a bit more efficient in cases
where the number of ROIs is quite large. Makes better use of the cluster
as well. Not to mention removes a call to `par_norm_layer` that uses the
old ipyparallel infrastructure.