Drop construction of some ipyparallel closures
These ipyparallel closures were used to compute various quantities for
large arrays in parallel using the cluster. Though they are no longer
needed as the same things are trivially computed with dask, which has
replaced this functionality in the workflow. Hence drop the definitions
of these closures as they are now unneeded.