Drop various unused imports
Over time we have changed many things about how the workflow works
including what functionality is needed in it. At times we have need to
import some things with `ipyparallel` to ensure that it will work when
we called it on workers. However this has resulted in a lot of leftover
imports over time that are no longer strictly needed. This cleans that
cruft out to simplify the workflow a bit.