Drop changing working directory on workers
As we are now using DRMAA and the default case uses the current working
directory for all job submitted, we can be confident that workers are
already in the current working directory used by the host where the
notebook is being run. This should hopefully mean an end to situations
where the workers somehow ended up being unable to find a file due to
workers being in a different directory.