Store shifts from registration
As one is likely curious how the registration algorithm translated the
frames, we store the shifts for later inspection. The shifts are
basically free to store as we have already computed them and persisted
that result in order to translate the frames. So this simply amounts to
a rechunk and store operation.