Drop zipping of Zarr groups
Initially the thought was that we would be making more use of the
intermediate files and possibly sharing them. Hence we zipped up the
Zarr groups into a single file to making them easier to distribute. In
practice only the final results are what we seem to share and
distribute. Those final results are also placed in HDF5 to make it easy
for users to pull the data into their favorite language easily.

Given this change, it makes more sense to not zip up the Zarr groups,
but leave them as directories on the file system. This avoids spending
time zipping the Zarr groups up, which can only be done in a single
process. Thus keeping IO parallel for reading and writing. Also if there
was any remaining concern about resource usage on the notebook process,
dropping the zipping step eliminates the last big consumer.

Not to mention Zarr's `DirectoryStore`s are a bit nicer to use as there
is no need to close them afterwards. All files with the `DirectoryStore`
are opened once for reading and closed immediately after the data is
read. This works nicely in place of any other stand-ins we have like
`LazyZarrDataset`s by already providing the behavior we would want
making them perfect for data viewers at the end of each workflow step.